bug fix for firms not adding children correctly
issue summary:

main.js depends on firm.js and firms.js  (needs the firm types for getNewFirmOfType, and needs Firm to override the Firm.addChildFirm function because it depends on getNewFirmOfType, which depends on firms, which depends on firm, so it cannot be defined in firm.js. Firm.addChildFirm is defined as a placeholder method and then overrriden in main.js, however, firms.js was already created in memory, so even though Firm had its mthod updated and firms inherits from Firm, firms had inherited the outdated function. Forcing them to inherit (`Mine.addChildFirm = Firm.addChildFirm;`) after defining the `Firm.addChildFirm` function and before new fims are created does the trick. Another bug fix is to do the following:

```
case 'forester': {
  tmp = new Forester(sellPrice);
  tmp.addChildFirm = Firm.addChildFirm;
  return tmp;
}
```

and set the function explicitly each time a firm is created, but that seemed inefficient and was changed to simply set the method pointer in `start` one time.