bug fix for addChildFirm
Even when updating the Mine, Smith addChildFirm methods, then creating a new Mine afterwards, the new mine would still have the old addChildFirm. We now set the new Mine's addChildFirm explicitly in getNewFirmOfType before returning. This revealed another bug: Firm.addChildFirm being overriden in main doesn't have access to `this`, so in Firm under doExpansion (where addChildFirm is called), we pass in the firm's type (this.type()) and sell price, so we have access to them and don't need `this` in main.js when setting the updated addChildFirm method.